+++ CHANGELOG.md
@@ -8,7 +8,9 @@ To know more about breaking changes, see the [Migration Guide][].
 
 ## Unreleased
 
-*None.*
+**Fixes**
+
+- Fix iOS/macOS returning all media types when `filterOption` is not specified instead of respecting the requested type.
 
 ## 3.8.2