+++ CHANGELOG.md
@@ -11,6 +11,7 @@ To know more about breaking changes, see the [Migration Guide][].
 **Fixes**
 
 - Fix iOS/macOS returning all media types when `filterOption` is not specified instead of respecting the requested type.
+- Fix Android returning all media types when `filterOption` is not specified instead of respecting the requested type.
 
 ## 3.8.2