+++ CHANGELOG.md
@@ -8,6 +8,10 @@ To know more about breaking changes, see the [Migration Guide][].
 
 ## Unreleased
 
+*None.*
+
+## 3.8.3
+
 **Fixes**
 
 - Fix all media assets being returned when `filterOption` is not specified instead of respecting the requested type.